refactor: move fixture in utils module
